ref: change structure
