+++ server.js
@@ -3,7 +3,7 @@
 const fastify = require('fastify')({ logger: true });
 const path = require('node:path');
 const { createDatabase } = require('./src/database/connection');
-const { errorHandler } = require('./src/controllers/error-handler');
+const { errorHandler } = require('./src/error-handler');
 
 const result = require('dotenv').config();
 if (result.error) {
