+++ src/routes/players.js
@@ -1,10 +1,10 @@
 'use strict';
 
 const {
-  getPlayerHandler,
-  registerPlayerHandler,
-  deletePlayerHandler,
-} = require('../controllers/players');
+  findPlayerByUsername,
+  createPlayer,
+  deletePlayer,
+} = require('../database/player');
 
 const Player = {
   type: 'object',
@@ -56,9 +56,40 @@ const deletePlayerOpts = {
 };
 
 const playerRoutes = (fastify, options, done) => {
-  fastify.get('/players/:username', getPlayerOpts, getPlayerHandler);
-  fastify.post('/players/register', registerPlayerOpts, registerPlayerHandler);
-  fastify.delete('/players/:username', deletePlayerOpts, deletePlayerHandler);
+  fastify.get(
+    '/players/:username',
+    getPlayerOpts,
+    (req, reply) => {
+      const { username } = req.params;
+      const player = findPlayerByUsername(username);
+      reply.send(player);
+    },
+  );
+
+
+  fastify.post(
+    '/players/register',
+    registerPlayerOpts,
+    (req, reply) => {
+      const { username, hashedPassword } = req.body;
+
+      createPlayer(username, hashedPassword);
+      const user = findPlayerByUsername(username);
+
+      reply.code(201).send(user);
+    },
+  );
+
+  fastify.delete(
+    '/players/:username',
+    deletePlayerOpts,
+    (req, reply) => {
+      const { username } = req.params;
+      deletePlayer(username);
+
+      reply.send({ 'message': `Player with ${username} has been deleted` });
+    },
+  );
 
   done();
 };
